Pre-assign the most restricted lands/orbs.
This introduces some bias into the randomization. In an unbiased
randomization (meaning every valid permutation has equal
probability), Orb of Winter is assigned to Volcanic Wasteland
most of the time.

We don't like this, so we assign Volcanic Wasteland first, giving
it close to a 25% chance of each orb it can have. We do a similar
thing for other restricted lands/orbs because it should make
things seem more random while being faster.