Account for lands where Phasing doesn't work well.
We can't use Phasing if there's nothing to phase through. So if the
native orb is Impact, and we add Phasing as a guest, that's not
going to make Impact useful. Therefore, do not use Phasing to combo
with those orbs that need it in such lands. (However, Phasing can
still be a random secondary orb.)

If Phasing is the native orb, just add a chance to combo it like
others, regardless of the land it appears in.